backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
django-cors-headers==4.3.1

# WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# HTTP/Async
aiohttp==3.9.1
requests==2.31.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Security
argon2-cffi==23.1.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
